requirements.txt
torch
torchvision
torchaudio
transformers>=4.37.0
datasets
aiohttp
streamlit
numpy
Pillow
tqdm
accelerate
sentencepiece
pycocoevalcap
gradio>=3.50
